Add OLS action to pod Actions dropdown using redux
Adds an "Ask OpenShift Lightspeed" action to the Actions dropdown on the
pod details page and to the kebab menu of the pods list page.

The callback is just a placeholder for now to show that the integration
works.

Adds a redux store for the plugin, which allows passing data from the
pod page to the OpenShift Lightspeed page.