Add popover UI for OLS
Instead of having a separate page for OLS, switch to using a popover UI.
When the web console loads, add the popover button in the bottom right.
The popover UI can be opened either by clicking on the button in the
bottom right or by triggering one of the "Ask OpenShift Lightspeed"
actions on resource pages.

Also adds a "OpenShift Lightspeed version" entry to the "Details" card
of the Home > Overview dashboard.

Also, when chatting about a k8s resource, show a dynamically updating
status indicator for that resource. (Currently only supports Job
resources.)