Implement separate user feedback for each response and move to redux
Store the user feedback in redux as part of the chat history so that we
can store different user feedback for each response and persist it when
the user feedback area is closed and when the whole OLS popover UI is
closed.

Also converts the chat history in Redux from a standard Array to an
Immutable List so that we can use the Immutable.js getIn function to
extract specific fields from within the nested data.