Change authorization check logic to always show UI
Instead of hiding the popover button when the authorization check fails,
show the UI with an error message.

If the `authorized` endpoint returns a 401 or 403 error, display a
message explaining that the user doesn't have permission to use
OpenShift Lightspeed and hide the prompt input. Otherwise, if the
`authorized` call succeeds or if it return an error other than 401 or
403, show the UI as usual without a message.