Switch from "injecting" YAML into prompt to "attaching" YAML to prompt
Add "Attach context" menu that allows attaching either a resource's full
YAML (with metadata.managedFields omitted) or just the status section of
its YAML.

The menu is shown for CronJobs, DaemonSets, Deployments, Jobs, Pods,
ReplicaSets and StatefulSets.

When the prompt is submitted, the attached context is appended to the
prompt text.

The YAML reflects the data when the context was attached. So if there is
a change in the resource's YAML data between adding the attachment to
the prompt and submitting the prompt, that change will not be reflected
in the submitted prompt.