Add the ability to attach container logs to the prompt
Uses a modal to allow you to select which container's logs you want and
how many lines from the log you want to include.

Only implemented for the pod details page for now.

Also removes the checkmarks from the attach menu. They don't make sense
for the Logs case since you can add multiple logs for a single pod.

Also moves the attach menu tooltip to above the menu button and hides it
when the menu is open.

Also increases the popover's width.