Pass attachments in separate fields to OLS
So that OLS has enough context to know what each attachment belongs to,
we now also include the resource's `kind` and `metadata` with the
`status` YAML attachment.

Similarly, we previx log attachments with a line of text explaning which
container and pod it belongs to.

Also removes the attachment's `options` data from redux since it is no
longer used.