Remove tooltip from Sliders and add current count to labels instead
For the attach logs modal and the attach events modal, remove the
tooltip above the Slider component. Instead show the currently selected
count in the input label. This solves a problem where the tooltip would
jitter as the slider was dragged. It also means the the count is always
visible without having to hover over the slider.