Use multicolor logo in light theme and monocolor in dark theme
When light theme is enabled, use the multicolor version of the OpenShift
Lightspeed logo everywhere and never use the monocolor logo.

When dark theme is enabled, use the dark monocolor version of the logo
everywhere and never use the multicolor logo.

The multicolor logo will be used in the operator catalog even with dark
theme, but that is not controlled by this plugin.