Fix attach events and attach logs modals to call onClose on unmount
Because these modals didn't call `onClose` when the components unmounted
(e.g.  when you hit the browser back button to leave the page), the
`isOpen` flag continued to be `true`. That meant that if you later
returned to a page with the same modal (another pod details page), the
modal would unexpectedly open.

Also makes sure we clear the attach events modal's `error` and
`isLoading` state if the resource context changes. Currently, this
shouldn't matter since changing the resource context only happens when
you navigate to a different page, which would unmount the modal and
clear the state anyway. However, it might be possible in the future.